feat: use cwd (#26)
issues:
* feat: use cwd (#25)

commits:
* feat: use cwd (700cf12)
