+++ src/index.ts
@@ -1,23 +1,20 @@
-import { validateSkippedStacks } from '@mnrendra/obtain-tsconfig-paths'
-
 import async from './async'
 import sync from './sync'
 
 export type {
   Aliases,
-  Alias,
+  Alias
+} from '@mnrendra/types-aliases'
+
+export type {
   TSConfig,
   CompilerOptions,
   BaseURL,
   Paths,
-  TSConfigPaths,
-  Options,
-  SkippedStacks,
-  ValidSkippedStacks
-} from './types'
+  TSConfigPaths
+} from '@mnrendra/obtain-tsconfig-paths'
 
 export {
   async as parseTSConfigAlias,
-  sync as parseTSConfigAliasSync,
-  validateSkippedStacks
+  sync as parseTSConfigAliasSync
 }
