--- src/types/index.ts
@@ -1,20 +1,0 @@
-import type Options from './Options'
-
-export type {
-  Aliases,
-  Alias
-} from '@mnrendra/types-aliases'
-
-export type {
-  TSConfig,
-  CompilerOptions,
-  BaseURL,
-  Paths,
-  TSConfigPaths,
-  SkippedStacks,
-  ValidSkippedStacks
-} from '@mnrendra/obtain-tsconfig-paths'
-
-export type {
-  Options
-}
