+++ tests/dummies/index.ts
@@ -1,7 +1,5 @@
 import defaultValue from './defaultValue'
-import tsConfigValues from './tsConfigValues'
 
 export {
-  defaultValue,
-  tsConfigValues
+  defaultValue
 }
