+++ tests/index.ts
@@ -1,9 +1,5 @@
 import * as dummies from './dummies'
-import * as mocks from './mocks'
-import * as unmocks from './unmocks'
 
 export {
-  dummies,
-  mocks,
-  unmocks
+  dummies
 }
