--- tests/mocks/index.ts
@@ -1,7 +1,0 @@
-import mockedReadAsync from './readAsync'
-import mockedReadSync from './readSync'
-
-export {
-  mockedReadAsync,
-  mockedReadSync
-}
