--- tests/unmocks/index.ts
@@ -1,7 +1,0 @@
-import unmockReadAsync from './readAsync'
-import unmockReadSync from './readSync'
-
-export {
-  unmockReadAsync,
-  unmockReadSync
-}
